GeoSpecc/geospec_platform/backend/requirements.txt
flask
flask-cors
tensorflow
numpy
opencv-python
geopandas
shapely
rasterio
scikit-learn
matplotlib
pillow